+++ src/pulumi/core-dtap/analytics/__init__.py
@@ -6,3 +6,4 @@ platform_outputs["analytics"] = {}
 from . import datafactory
 from . import databricks
 from . import dbt
+from . import jupyterlab