+++ src/pulumi/core-shared/analytics/shared_kubernetes_cluster.py
@@ -1,28 +1,53 @@
+import platform
+from pulumi import ResourceOptions
 import pulumi_random
 from pulumi_azure_native import containerservice
 
 from ingenii_azure_data_platform.iam import GroupRoleAssignment
 from ingenii_azure_data_platform.utils import generate_resource_name, lock_resource
 
-from management import resource_groups, user_groups
+from management import resource_groups, resource_group_outputs, user_groups
 from network.vnet import hosted_services_subnet
-from project_config import platform_config, platform_outputs
-
-runtime_config = platform_config["analytics_services"]["datafactory"][
-    "integrated_self_hosted_runtime"
-]
+from project_config import azure_client, platform_config, platform_outputs
 
 # ----------------------------------------------------------------------------------------------------------------------
-# SHARED KUBERNETES CLUSTER
+# SHARED KUBERNETES CLUSTER -> BASE CONFIGURATIONS
 # ----------------------------------------------------------------------------------------------------------------------
 
 cluster_resource_name = "shared_cluster"
-resource_group_name = resource_groups["infra"].name
+cluster_resource_group_name = resource_groups["infra"].name
+configs = [
+    {
+        "config": platform_config["analytics_services"]["datafactory"]["integrated_self_hosted_runtime"],
+        "os": containerservice.OSType.WINDOWS,
+    },
+    {
+        "config": platform_config["analytics_services"]["jupyterlab"],
+        "os": containerservice.OSType.LINUX,
+    },
+]
+outputs = platform_outputs["analytics"]["shared_kubernetes_cluster"] = {}
 
-if runtime_config["enabled"]:
-    outputs = platform_outputs["analytics"]["shared_kubernetes_cluster"] = {}
+# Only create if a system requires it
+if any(config["config"]["enabled"] for config in configs):
+    
+    # ----------------------------------------------------------------------------------------------------------------------
+    # SHARED KUBERNETES CLUSTER -> NODE RESOURCE GROUP NAME
+    # ----------------------------------------------------------------------------------------------------------------------
 
-    admin_password = pulumi_random.RandomPassword(
+    # Resource group for Kubernetes nodes, otherwise Azure will create one
+    resource_group_config = platform_config["shared_kubernetes_cluster"]["resource_group"]
+    node_resource_group_name = generate_resource_name(
+            resource_type="resource_group",
+            resource_name=resource_group_config["display_name"],
+            platform_config=platform_config,
+        )
+
+    # ----------------------------------------------------------------------------------------------------------------------
+    # SHARED KUBERNETES CLUSTER -> CLUSTER
+    # ----------------------------------------------------------------------------------------------------------------------
+
+    windows_admin_password = pulumi_random.RandomPassword(
         resource_name=generate_resource_name(
             resource_type="random_password",
             resource_name=cluster_resource_name,
@@ -36,7 +61,6 @@ if runtime_config["enabled"]:
         override_special="_%@",
     )
 
-    # Note: Windows pool names must be 6 characters or fewer: https://docs.microsoft.com/en-us/azure/aks/windows-container-cli#limitations
     kubernetes_cluster = containerservice.ManagedCluster(
         resource_name=generate_resource_name(
             resource_type="kubernetes_cluster",
@@ -49,31 +73,18 @@ if runtime_config["enabled"]:
             managed=True,
         ),
         agent_pool_profiles=[
+            # At minimum, the cluster requires a system Linux pool
             containerservice.ManagedClusterAgentPoolProfileArgs(
                 availability_zones=["1"],
                 count=1,
                 enable_auto_scaling=True,
                 max_count=1,
                 min_count=1,
-                mode="System",
+                mode=containerservice.AgentPoolMode.SYSTEM,
                 name="systempool",
-                node_labels={"OS": "Linux"},
+                node_labels={"OS": containerservice.OSType.LINUX},
                 os_type=containerservice.OSType.LINUX,
-                type="VirtualMachineScaleSets",
-                vm_size="Standard_B2ms",
-                vnet_subnet_id=hosted_services_subnet.id,
-            ),
-            containerservice.ManagedClusterAgentPoolProfileArgs(
-                availability_zones=["1"],
-                count=1,
-                enable_auto_scaling=True,
-                max_count=1,
-                min_count=1,
-                mode="User",
-                name="win1",
-                node_labels={"OS": "Windows"},
-                os_type=containerservice.OSType.WINDOWS,
-                type="VirtualMachineScaleSets",
+                type=containerservice.AgentPoolType.VIRTUAL_MACHINE_SCALE_SETS,
                 vm_size="Standard_B2ms",
                 vnet_subnet_id=hosted_services_subnet.id,
             ),
@@ -93,14 +104,16 @@ if runtime_config["enabled"]:
             network_plugin=containerservice.NetworkPlugin.AZURE,
             network_policy=containerservice.NetworkPolicy.AZURE,
         ),
-        resource_group_name=resource_group_name,
+        node_resource_group=node_resource_group_name,
+        opts=ResourceOptions(delete_before_replace=True),
+        resource_group_name=cluster_resource_group_name,
         sku=containerservice.ManagedClusterSKUArgs(
             name=containerservice.ManagedClusterSKUName.BASIC,
             tier=containerservice.ManagedClusterSKUTier.FREE,
         ),
         tags=platform_config.tags,
         windows_profile=containerservice.ManagedClusterWindowsProfileArgs(
-            admin_password=admin_password.result,
+            admin_password=windows_admin_password.result,
             admin_username="runtimeclusteradmin",
         ),
     )
@@ -111,15 +124,20 @@ if runtime_config["enabled"]:
     #    identity_profile: Optional[Mapping[str, ManagedClusterPropertiesIdentityProfileArgs]] = None,
     #    private_link_resources: Optional[Sequence[PrivateLinkResourceArgs]] = None,
 
-    for attr in ["fqdn", "id", "name"]:
-        outputs[attr] = getattr(kubernetes_cluster, attr)
-    outputs["resource_group_name"] = resource_group_name.apply(lambda name: name)
+    outputs.update({
+        "cluster_resource_group_name": cluster_resource_group_name.apply(lambda name: name),
+        "fqdn": kubernetes_cluster.fqdn,
+        "id": kubernetes_cluster.id,
+        "name": kubernetes_cluster.name,
+        "node_resource_group_name": node_resource_group_name,
+        "principal_id": kubernetes_cluster.identity.principal_id,
+    })
 
     # ----------------------------------------------------------------------------------------------------------------------
     # SHARED KUBERNETES CLUSTER -> ROLE ASSIGNMENTS
     # ----------------------------------------------------------------------------------------------------------------------
 
-    for assignment in runtime_config.get("iam", {}).get("role_assignments", []):
+    for assignment in platform_config["shared_kubernetes_cluster"]["cluster"]["iam"]["role_assignments"]:
         # User Group Assignment
         user_group_ref_key = assignment.get("user_group_ref_key")
         if user_group_ref_key is not None:
@@ -130,3 +148,102 @@ if runtime_config["enabled"]:
                 scope=kubernetes_cluster.id,
                 scope_description="kubernetes-cluster",
             )
+
+    # ----------------------------------------------------------------------------------------------------------------------
+    # SHARED KUBERNETES CLUSTER -> NODE RESOURCE GROUP ASSIGNMENTS
+    # ----------------------------------------------------------------------------------------------------------------------
+
+    node_resource_group_id = kubernetes_cluster.node_resource_group.apply(
+        lambda rg_name: f"/subscriptions/{azure_client.subscription_id}/resourceGroups/{rg_name}"
+    )
+
+    # Export resource group metadata
+    resource_group_outputs["kubernetes"] = {
+        "name": node_resource_group_name,
+        "location": kubernetes_cluster.location,
+        "id": node_resource_group_id,
+    }
+    for assignment in resource_group_config["iam"]["role_assignments"]:
+        # User Group Assignment
+        user_group_ref_key = assignment.get("user_group_ref_key")
+        if user_group_ref_key is not None:
+            GroupRoleAssignment(
+                principal_id=user_groups[user_group_ref_key]["object_id"],
+                principal_name=user_group_ref_key,
+                role_name=assignment["role_definition_name"],
+                scope=node_resource_group_id,
+                scope_description="kubernetes-cluster-node-resource-group",
+            )
+
+    # ----------------------------------------------------------------------------------------------------------------------
+    # SHARED KUBERNETES CLUSTER -> AGENT POOLS
+    # ----------------------------------------------------------------------------------------------------------------------
+
+    for idx, pool in enumerate(platform_config["shared_kubernetes_cluster"]["cluster"].get("linux_agent_pools", [])):
+        agent_pool_name = pool.get("name", f"linux{idx}")
+        containerservice.AgentPool(
+            resource_name=generate_resource_name(
+                resource_type="kubernetes_agent_pool",
+                resource_name=agent_pool_name,
+                platform_config=platform_config,
+            ),
+            agent_pool_name=agent_pool_name,
+            availability_zones=pool.get("availability_zones", ["1"]),
+            count=pool.get("count", 1),
+            enable_auto_scaling=pool.get("auto_scaling", True),
+            max_count=pool.get("max_count", 1),
+            min_count=pool.get("min_count", 1),
+            mode=containerservice.AgentPoolMode.USER,
+            node_labels={
+                **pool.get("labels", {}),
+                "OS": containerservice.OSType.LINUX,
+            },
+            os_type=containerservice.OSType.LINUX,
+            resource_group_name=cluster_resource_group_name,
+            resource_name_=kubernetes_cluster.name,
+            tags=platform_config.tags,
+            type=containerservice.AgentPoolType.VIRTUAL_MACHINE_SCALE_SETS,
+            vm_size=pool.get("vm_size", "Standard_B2ms"),
+            vnet_subnet_id=hosted_services_subnet.id,
+        )
+
+    # Check if any of the enabled features need Windows machines
+    need_windows = any(
+        config["config"]["enabled"] and config["os"] == containerservice.OSType.WINDOWS
+        for config in configs
+    )
+    windows_pools = platform_config["shared_kubernetes_cluster"]["cluster"].get("windows_agent_pools", [])
+    if need_windows and not windows_pools:
+        windows_pools = [{
+            "labels": {"addedBy": "platform"},
+            "name": "win1"
+        }]
+
+    for idx, pool in enumerate(windows_pools):
+        # Note: Windows pool names must be 6 characters or fewer: https://docs.microsoft.com/en-us/azure/aks/windows-container-cli#limitations
+        agent_pool_name = pool.get("name", f"win{idx}")
+        containerservice.AgentPool(
+            resource_name=generate_resource_name(
+                resource_type="kubernetes_agent_pool",
+                resource_name=agent_pool_name,
+                platform_config=platform_config,
+            ),
+            agent_pool_name=agent_pool_name,
+            availability_zones=pool.get("availability_zones", ["1"]),
+            count=pool.get("count", 1),
+            enable_auto_scaling=pool.get("auto_scaling", True),
+            max_count=pool.get("max_count", 1),
+            min_count=pool.get("min_count", 1),
+            mode=containerservice.AgentPoolMode.USER,
+            node_labels={
+                **pool.get("labels", {}),
+                "OS": containerservice.OSType.WINDOWS,
+            },
+            os_type=containerservice.OSType.WINDOWS,
+            resource_group_name=cluster_resource_group_name,
+            resource_name_=kubernetes_cluster.name,
+            tags=platform_config.tags,
+            type=containerservice.AgentPoolType.VIRTUAL_MACHINE_SCALE_SETS,
+            vm_size=pool.get("vm_size", "Standard_B2ms"),
+            vnet_subnet_id=hosted_services_subnet.id,
+        )