+++ src/pulumi/core-shared/management/__init__.py
@@ -6,4 +6,4 @@ platform_outputs["management"] = {}
 # Load sub-modules
 from . import resource_providers
 from .user_groups import user_groups
-from .resource_groups import resource_groups
+from .resource_groups import resource_groups, outputs as resource_group_outputs