+++ src/python/ingenii_azure_data_platform/ingenii_azure_data_platform/utils.py
@@ -44,6 +44,7 @@ def generate_resource_name(
         "devops_repo": "adorepo",
         "devops_variable_group": "adovg",
         "dns_zone": "dz",
+        "kubernetes_agent_pool": "kap",
         "kubernetes_cluster": "kc",
         "log_analytics_workspace": "law",
         "metric_alert": "ma",